Start of packaging, making it suitable for pypi
