+++ .gitignore
@@ -1,2 +1,4 @@
 .DS_Store
 *.pyc
+ENV/
+*.egg-info/
