+++ spoofmac/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf8 -*-
+from spoofmac.util import *
+from spoofmac.interface import *
