+++ spoofmac/version.py
@@ -1,0 +1,2 @@
+# -*- coding: utf8 -*-
+__version__ = '1.0.0'
